open_mes/requirements.txt
Django==5.1.7
djangorestframework==3.15.1
django-environ==0.11.2
django-debug-toolbar==4.3.0
django-cors-headers==4.3.1
django-static-md5url==0.1
Pillow==10.3.0
uuid6==2024.1.12
python-dateutil==2.9.0.post0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0